Partial fix for stdin bug(s) in simple read/write loop.
Root Cause

  The test_params program reads from stdin via in ? x, which uses the kroc.keyboard.process. This process calls
  BX.read.keyboard in a while loop guarded by c.in = (MOSTNEG INT). There were two bugs preventing the keyboard from
  working:

  Bug 1: MOSTNEG INT constant was zero (primary cause)

  The I_MINT instruction generates MOSTNEG INT = 0x80000000 (32-bit). This value was stored in the constmap with
  c_val type int, which correctly held 0x80000000 = -2147483648. But when the constant was passed through
  compose_ins() to create an INS_MOVE instruction, the AArch64 INS_MOVE handler's aarch64_validate_register() check
  rejected it:

  return (reg != (int)0x80000000 && reg != REG_UNDEFINED);

  The check reg != (int)0x80000000 matched the MOSTNEG INT constant 0x80000000, treating it as an "invalid register."
   The entire INS_MOVE was skipped with // SKIPPED: invalid register constant, so c.in was never set to MOSTNEG INT.
  With c.in = 0 instead of MOSTNEG INT, the WHILE c.in = (MOSTNEG INT) condition was immediately false, the
  BX.read.keyboard call was never reached, and the keyboard process sent NUL bytes endlessly without blocking.

  Fix: Removed the overzealous aarch64_validate_register check from INS_MOVE. The check was incorrectly applied to
  ALL operands including ARG_CONST values, when it should only apply to register numbers.

  Bug 2: BX call parameter block offset (contributing issue)

  In compose_bcall_aarch64, the LEA displacement for the parameter block pointer was hardcoded as 4 (1 word on
  32-bit) instead of (1 << WSH) = 8 (1 word on 64-bit). Additionally, compose_aarch64_kcall was called with
  regs_in=2, causing it to overwrite cparam[0] (the parameter block pointer) with the transputer B-register value.

  Fix: Changed displacement to (1 << WSH) and regs_in from 2 to 1.

  Files changed

  - tools/tranx86/archaarch64.c: Removed aarch64_validate_register from INS_MOVE; fixed BX call parameter offset and
  kcall regs_in
  - tools/tranx86/etcrtl.c: Fixed I_MINT to use (int)0x80000000 (32-bit MOSTNEG)
  - tools/tranx86/tstack.h: Widened constmap_t.c_val to intptr_t
  - tools/tranx86/tstack.c: Updated constmap_regconst and constmap_otherlab return types to intptr_t

  The remaining range error during the BYTE c.in conversion is a separate issue to investigate next.

+++ tools/tranx86/tstack.c
@@ -325,11 +325,11 @@ int constmap_typeof (int reg)
 	return VALUE_UNDEFINED;
 }
 /*}}}*/
-/*{{{  int constmap_regconst (int reg)*/
+/*{{{  intptr_t constmap_regconst (int reg)*/
 /*
  *	returns constant value of `reg' -- errors if not here
  */
-int constmap_regconst (int reg)
+intptr_t constmap_regconst (int reg)
 {
 	int i;
 
@@ -371,11 +371,11 @@ void constmap_modregconst (int reg, int val)
 	exit (EXIT_FAILURE);
 }
 /*}}}*/
-/*{{{  int constmap_otherlab (int reg)*/
+/*{{{  intptr_t constmap_otherlab (int reg)*/
 /*
  *	returns the other label for VALUE_LABDIFF
  */
-int constmap_otherlab (int reg)
+intptr_t constmap_otherlab (int reg)
 {
 	int i;
